requirements.txt
openai==1.12.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pandas==2.1.4
Pillow==10.2.0
streamlit==1.31.0